+++ proxyctl/cmd/root.go
@@ -57,6 +57,7 @@ func init() {
 		toolsCmd,
 		portCmd,
 		profileCmd,
+		tunCmd,
 		versionCmd,
 	)
 }